Haze can prevent a Pokémon from attacking after curing freeze
If an enemy Pokémon freezes the player Pokémon while it must recharge from Hyper Beam and then cures the freeze status with Haze, the player will be unable to move for the rest of the battle. This is because, while the Hyper Beam recharge bit is reset when the player inflicts freeze, it is not reset when the enemy does so. Haze loads $ff as the target's last selected move if it is frozen, which prevents it from moving when the player hasn't selected a move. However, because the recharge bit is still set, the player is unable to select a new move, and because $ff is the last selected move, the routine that checks and resets the recharge bit is never run. This permanently locks the player out of making a move.