Cinnabar Island's left-facing shore tiles point to invalid Pokémon
Due to an oversight in the localizations of Red/Blue, the left-facing shore tiles of Cinnabar Island will allow players to encounter Pokémon from a glitched table rather than the normal table for the route. This oversight was responsible for the discovery of Missingno. as well as other various glitch Pokémon. This also fixes the below bug and various other 2x2 block encounter glitches, and was fixed in Yellow.