AI trainer HUD does not update when it uses healing items
When an AI trainer uses a Full Heal (such as Brock), the status of the opposing pokemon does not update until after the player's turn. When playing with Super Gameboy colors, and the opponent is in yellow or red health, and the AI trainer uses a healing item to restore HP up into a new color (such as back into green health), the color of the HP bar will not update until after the player's turn. The AI's version of using a Full Restore combines both of these observations. Why does this happen? The cause is that, unlike what is done in the core battle engine, the function DrawEnemyHUDAndHPBar never gets run when executing the AI item functions.